Rename to channels.host
