+++ src/Channels.Host/Startup.cs
@@ -1,17 +1,14 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using tanka.graphql.samples.Chat.Host.Schemas;
+using tanka.graphql.samples.channels.host.logic;
 using tanka.graphql.server;
 using tanka.graphql.tools;
 
-namespace tanka.graphql.samples.Chat.Host
+namespace tanka.graphql.samples.channels.host
 {
     public class Startup
     {
-        // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
             // add schema
@@ -20,9 +17,9 @@ namespace tanka.graphql.samples.Chat.Host
                 {
                     var schemaBuilder = SchemaLoader.Load();
 
-                    var chat = new Domain.Chat();
-                    var service = new ChatResolverService(chat);
-                    var resolvers = new ChatResolvers(service);
+                    var chat = new Channels();
+                    var service = new ResolverService(chat);
+                    var resolvers = new Resolvers(service);
 
                     var schema = SchemaTools.MakeExecutableSchemaWithIntrospection(
                         schemaBuilder,
