+++ src/Channels.Host/appsettings.json
