+++ src/Channels.Host/logic/SchemaLoader.cs
@@ -5,7 +5,7 @@ using System.Text;
 using tanka.graphql.sdl;
 using tanka.graphql.type;
 
-namespace tanka.graphql.samples.Chat.Host.Schemas
+namespace tanka.graphql.samples.channels.host.logic
 {
     public static class SchemaLoader
     {
@@ -14,7 +14,7 @@ namespace tanka.graphql.samples.Chat.Host.Schemas
             var idl = LoadIdlFromResource();
 
             var builder = new SchemaBuilder();
-            Sdl.Import(Parser.ParseDocument(idl), builder);
+            builder.Sdl(idl);
 
             return builder;
         }
@@ -27,7 +27,8 @@ namespace tanka.graphql.samples.Chat.Host.Schemas
         {
             var assembly = Assembly.GetExecutingAssembly();
             var resourceStream =
-                assembly.GetManifestResourceStream("tanka.graphql.samples.Chat.Host.Schemas.Chat.graphql");
+                assembly.GetManifestResourceStream("tanka.graphql.samples.channels.host.logic.Schema.graphql");
+
             using (var reader =
                 new StreamReader(resourceStream ?? throw new InvalidOperationException(), Encoding.UTF8))
             {
