--- src/Host.Logic/Host.Logic.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <RootNamespace>tanka.graphql.samples.Host.Logic</RootNamespace>
-    <AssemblyName>tanka.graphql.samples.Host.Logic</AssemblyName>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.1.0" />
-    <PackageReference Include="tanka.graphql" Version="0.5.0" />
-    <PackageReference Include="tanka.graphql.server" Version="0.5.0" />
-  </ItemGroup>
-
-
-</Project>
